calculation logic added
added logic to the calculator. to take the users input from the three
variables and calculate a number based on the selections.

addition subtraction multiply and divide all work as required.

added message when user tries to divide by 0

calculator will take the sum and apply it to numOne for the next calculation